HCSVLAB-1019: add handling of whitespace in name of uploaded files
The existing ingest code didn't appear to have any mechanisms in place to handle documents which have spaces in their file names. This is a small patch to enable the api to handle whitespace in filenames by subbing them with the '%20' character in the file reference URI. When the ingest looks for documents it now decodes the file path URI to properly decode the escaped whitespace before the file path is saved to the document record.