dev-support: update multipart test to reflect documented api call
The documented api call to add a document to an existing item via a multipart POST request is resulting in an error due to the Rack middleware not automatically parsing the metadata parameter contained within the multipart body from string to JSON. This change to the auto-test removes the explicit parsing to JSON, to reflect that the metadata currently reaches the controller as a string rather  a JSON object. As such the auto test now fails with the same error as produced by the bug in production.